The Ball Project Official - Noe
Lesson one project on using physics in unity and pretty much starting with the basics. Added a Free-cam Code made by Ashley Davis on GitHub. (((BE ADVISED))) Physics get quirky, they are set for now to use no changes in mass, but if it seems to move off a different place its most definitely because of that... you have been advised.